moved function to utility package
